Updated mluti-pages: /routes folder has been renamed into /pages, pages components have been CamelCased, and Links were changed into NavLinks to be able to pass the active state to the className
