+++ src/main.jsx
@@ -2,14 +2,15 @@ import React from "react";
 import ReactDOM from "react-dom/client";
 
 import { createBrowserRouter, RouterProvider } from "react-router-dom";
-import Home from "./routes/home";
-import Planner from "./routes/planner";
-import Recipes from "./routes/recipes";
-import Pantry from "./routes/pantry";
-import Groceries from "./routes/groceries";
-import Login from "./routes/login";
-import Register from "./routes/register";
-import ErrorPage from "./routes/error-page";
+
+import Home from "./pages/Home";
+import Planner from "./pages/Planner";
+import Recipes from "./pages/Recipes";
+import Pantry from "./pages/Pantry";
+import Groceries from "./pages/Groceries";
+import Login from "./pages/Login";
+import Register from "./pages/Register";
+import NotFound from "./pages/NotFound";
 
 import "./index.css";
 
@@ -17,7 +18,7 @@ const router = createBrowserRouter([
   {
     path: "/",
     element: <Home />,
-    errorElement: <ErrorPage />,
+    errorElement: <NotFound />,
   },
   {
     path: "/planner",
