+++ src/pages/Pantry.jsx
